Moved plot_umaps() to utils.py
