+++ src/cellxgene_scvi.py
@@ -9,7 +9,6 @@
 
 # %%
 # Standard python libraries
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
@@ -21,7 +20,7 @@ import scanpy as sc
 import scvi
 
 # Local utilities
-from utils import load_config, get_data_dir, initialize_output_directory, save_package_versions, compare_cell_metadata_cols
+from utils import load_config, get_data_dir, initialize_output_directory, save_package_versions, compare_cell_metadata_cols, plot_umaps
 
 # - suppress "Transforming to str index" warnings when loading CellxGene census 
 #   data into anndata
@@ -173,70 +172,6 @@ def compute_frac_correct(
         predicted_cell_types[mask] == gold_standard_cell_types[mask]
     ).mean()
     return frac_correct
-
-# Plot a set of UMAPs with consistent figure dimensions and spacing.
-def plot_umaps(adata, out_dir, plot_specs, fixed_dpi=300, plot_area_in=5.2, left_margin_in=0.55, bottom_margin_in=0.55, top_margin_in=0.35, right_margin_continuous_in=2, right_margin_categorical_in=4):
-    """
-    Plot UMAPs with consistent figure dimensions and spacing.
-    
-    Parameters
-    ----------
-    adata : AnnData
-        AnnData object containing the data to plot.
-    out_dir : str
-        Directory to save the output plots.
-    plot_specs : list of tuples
-        List of (color_col, filename, title) tuples specifying what to plot.
-    fixed_dpi : int, default 300
-        DPI for saved figures.
-    plot_area_in : float, default 5.2
-        Width and height of the plot area in inches.
-    left_margin_in : float, default 0.55
-        Left margin in inches.
-    bottom_margin_in : float, default 0.55
-        Bottom margin in inches.
-    top_margin_in : float, default 0.35
-        Top margin in inches.
-    right_margin_continuous_in : float, default 2
-        Right margin in inches for continuous data legends.
-    right_margin_categorical_in : float, default 4
-        Right margin in inches for categorical legends.
-    """
-
-    for color_col, filename, title in plot_specs:
-        series = adata.obs[color_col]
-        is_continuous = pd.api.types.is_numeric_dtype(series)
-
-        right_margin_in = (
-            right_margin_continuous_in if is_continuous else right_margin_categorical_in
-        )
-
-        fig_w = left_margin_in + plot_area_in + right_margin_in
-        fig_h = bottom_margin_in + plot_area_in + top_margin_in
-        fig, ax = plt.subplots(figsize=(fig_w, fig_h), dpi=fixed_dpi)
-
-        # Position axis explicitly so the plotting area is always the same size
-        ax_left = left_margin_in / fig_w
-        ax_bottom = bottom_margin_in / fig_h
-        ax_width = plot_area_in / fig_w
-        ax_height = plot_area_in / fig_h
-
-        with plt.rc_context():
-            sc.pl.umap(
-                adata,
-                color=color_col,
-                title=title,
-                ax=ax,
-                show=False,
-                legend_loc="right margin",
-                colorbar_loc="right",
-            )
-
-            ax.set_position([ax_left, ax_bottom, ax_width, ax_height])
-            ax.set_box_aspect(1)
-
-            fig.savefig(os.path.join(out_dir, filename), dpi=fixed_dpi, bbox_inches="tight", pad_inches=0.02)
-            plt.close(fig)
 
 # %% [markdown]
 # # Main script
